dom/file/tests/test_fileapi_twice_worker.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for FileReader API in workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="common_fileReader.js"></script>
</head>

<body>
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.requestLongerTimeout(3);

test_setup()
.then(data => {
  let worker = new Worker('worker_fileReader.js');
  worker.postMessage({ tests: 'twice', data });

  worker.onmessage = event => {
    if (event.data.type == 'finish') {
      SimpleTest.finish();
      return;
    }

    if (event.data.type == 'check') {
      ok(event.data.status, event.data.msg);
      return;
    }

    ok(false, "Unknown message.");
  }
});

</script>
</body>
</html>
